Tests

+++ backend/Authenticatie/Services/GebruikerService.cs
@@ -12,7 +12,7 @@ public class GebruikerService : IGebruikerService{
         _emailService = emailService;
     }
     public async Task<string> Registreer(string voornaam, string achternaam, string email, string wachtwoord, VerificatieToken verificatieToken, GebruikerContext context){
-        Klant klant = new Klant(voornaam, achternaam, email, wachtwoord){VerificatieToken = verificatieToken};
+        Klant klant = new Klant(voornaam, achternaam, email, wachtwoord){VerificatieToken = verificatieToken, TokenId = verificatieToken.Token};
         if(await CheckDomainIsDisposable(email)) return "DisposableMailError"; 
         else if(context.Klanten.Any(k => k.Email == email)) return "EmailInUseError"; 
         if(await context.Rollen.FirstOrDefaultAsync(r => r.Naam == "Klant") != null){
@@ -22,7 +22,7 @@ public class GebruikerService : IGebruikerService{
         }
         await context.Klanten.AddAsync(klant);
         await context.SaveChangesAsync();
-        await _emailService.Send(email, klant.VerificatieToken.Token); //Hier nog juiste content toevoegen
+        await _emailService.Send(email, "http://localhost:3000/verify?token=" + klant.VerificatieToken.Token + "&email="+ email, "Email Verification"); //Hier nog juiste content toevoegen
         return "Success";
     }
     public async Task<string> Login(string email, string wachtwoord, GebruikerContext context){
@@ -81,6 +81,7 @@ public class GebruikerService : IGebruikerService{
         return tuple;
     }
     public async Task<string> InitiatePasswordReset(Klant klant, GebruikerContext context){
+        if(klant == null) return "Error";
         if(klant.AuthenticatieTokenId != null){
             AuthenticatieToken authenticatieToken = context.AuthenticatieTokens.First(a => a.Token == klant.AuthenticatieTokenId);
             context.AuthenticatieTokens.Remove(authenticatieToken);
@@ -88,13 +89,13 @@ public class GebruikerService : IGebruikerService{
         }
         klant.AuthenticatieToken = new AuthenticatieToken(){Token = Guid.NewGuid().ToString(), VerloopDatum = DateTime.Now.AddDays(1)};
         await context.SaveChangesAsync();
-        await _emailService.Send(klant.Email, klant.AuthenticatieTokenId!);
+        await _emailService.Send(klant.Email, klant.AuthenticatieTokenId!, "Password Reset");
         return "Success";
     }
     public async Task<string> ResetPassword(Klant klant, string token, string wachtwoord, GebruikerContext context){
         AuthenticatieToken authenticatieToken = context.AuthenticatieTokens.FirstOrDefault(a => a.Token == token);
-        if(authenticatieToken == null || authenticatieToken.VerloopDatum < DateTime.Now) return "Error"; //Goede error message
-        if(klant.AuthenticatieTokenId != authenticatieToken.Token) return "Token matcht niet error"; //Goede error message
+        if(authenticatieToken == null || authenticatieToken.VerloopDatum < DateTime.Now) return "ExpiredTokenError"; //Goede error message
+        if(klant.AuthenticatieTokenId != authenticatieToken.Token) return "InvalidTokenError"; //Goede error message
         klant.AuthenticatieToken = null;
         klant.AuthenticatieTokenId = null;
         klant.Wachtwoord = wachtwoord;
